[Search Scraping - Web Service MVP]
Added a web service project to act as the middleware handling search requests.

Created a class library that will handle the scraping implementation details. Allows for the web service to be optional and simply transitioned a desktop client.

Fixed typo for 'scrape'.

Tests will be needed to test 'header spoofed' search responses and confirm regex functionality.