[UI - Search Result Banners]
Created a banner like control to be used to display results in a list box with a clickable hyperlink.

Backend needs to HTML decode the urls that are captured to make them immediately clickable in the frontend UI.

Prepared an extra test for handling bad captures.

Fixed the regex template to be able to not accidentally trigger re-capturing duplicate links (such as results with clickable thumbnails).