[Search Query Tool UI - MVVM Prototyping]
Refactored the code to work more along the lines of MVVM design.

Created a viewmodel project to contain the viewmodels to be displayed in the UI.

Deleted and cleaned up unnecessary files.

With MVVM being used, backing code in some of the controls has been removed(except for opening the banner hyperlink).

Cleaned up the XAML files to remove unnecessary naming and namespaces.