[Benchmark Test - Fallback Regex Parser]
Added some benchmarks to check the impact of the fallback search regex parser class. GC pressure is increased a fair amount.

Added a timeout to the Regex instance made by the search result capture class to ensure the regex pattern cannot be used for a DDOS. Also added the compile flag for the regex object.